refactor(bindings): extract the neural/VAE models into src/python/neural.rs (#385)
Third incremental split of src/python/mod.rs. Move the neural / VAE topic models
— ETM, DETM, InfoCTM, ProdLDA, and its contextualized variants CombinedTM and
ZeroShotTM — into a new src/python/neural.rs. These six share the AVITM
prior/option binding helpers (prior_from_str, build_avitm_options, the
encoder-input mode codec, parse_doc_embeddings, the default_* serde defaults),
which are used only by this group and move with them. FASTopic stays in mod.rs:
it is optimal-transport and uses none of these helpers.

Shared cross-model helpers, type aliases, and save-format tags stay in mod.rs,
reached via `use super::*`. No public API change: all six classes are still
registered in the same #[pymodule] fn (`use neural::{...}`) and imported as
topica.ETM / topica.ProdLDA / etc. Update the #381 IMPL map bindings to point at
the new module and regenerate the model map.

mod.rs: 16414 -> 13555 lines (18407 at the start of #385). Gates green: cargo
check (default + python), clippy -D warnings, fmt, preflight, and the full pytest
suite (3321 passed, incl. all six models' gold + save/load + conformance +
registry).
