+++ python/topica/registry.py
@@ -255,7 +255,7 @@ def _i(source: str, binding: str, core: str, feature: str, validation: str) -> I
 IMPL: dict[str, ImplInfo] = {
     "LDA": _i("src/model.rs", "src/python/mod.rs", "SparseLDA collapsed Gibbs (model.rs, sampler.rs)", "", "parity/lda_gold.py, parity/mallet_parity.py"),
     "CTM": _i("topica-core/src/ctm.rs", "src/python/mod.rs", "CTM/STM variational core (topica-core)", "", "parity/ctm_gold.py, parity/ctm_r_compare.py"),
-    "ProdLDA": _i("src/prodlda.rs", "src/python/mod.rs", "hand-coded batched VAE (prodlda.rs)", "", "parity/prodlda_gold.py, parity/prodlda_compare.py"),
+    "ProdLDA": _i("src/prodlda.rs", "src/python/neural.rs", "hand-coded batched VAE (prodlda.rs)", "", "parity/prodlda_gold.py, parity/prodlda_compare.py"),
     "HDP": _i("src/hdp.rs", "src/python/mod.rs", "collapsed Gibbs (model.rs, sampler.rs)", "", "parity/hdp_gold.py"),
     "NMF": _i("src/nmf.rs", "src/python/nmf_lsa.rs", "multiplicative-update matrix factorization", "", "parity/nmf_vs_sklearn.py"),
     "LSA": _i("src/lsa.rs", "src/python/nmf_lsa.rs", "truncated SVD (linalg)", "", "parity/lsa_vs_sklearn.py"),
@@ -278,12 +278,12 @@ IMPL: dict[str, ImplInfo] = {
     "PT": _i("src/pt.rs", "src/python/mod.rs", "collapsed Gibbs over pseudo-documents", "", "parity/pt_gold.py"),
     "BTM": _i("src/btm.rs", "src/python/btm.rs", "collapsed Gibbs over biterms", "", "parity/btm_compare.py, tests/test_btm.py"),
     "DTM": _i("src/dtm.rs", "src/python/mod.rs", "variational Kalman over time slices", "", "parity/dtm_gold.py"),
-    "DETM": _i("src/detm.rs", "src/python/mod.rs", "embedding VAE + LSTM q(eta) (etm_vae.rs)", "", "parity/detm_gold.py"),
+    "DETM": _i("src/detm.rs", "src/python/neural.rs", "embedding VAE + LSTM q(eta) (etm_vae.rs)", "", "parity/detm_gold.py"),
     "HLDA": _i("src/hlda.rs", "src/python/hierarchical.rs", "nested-CRP collapsed Gibbs", "", "parity/hlda_gold.py"),
     "PA": _i("src/pa.rs", "src/python/hierarchical.rs", "collapsed Gibbs over a topic DAG", "", "parity/pa_gold.py"),
     "BERTopic": _i("src/bertopic.rs", "src/python/embedding_cluster.rs", "embedding clustering (cluster.rs, reduce.rs, represent.rs)", "embeddings", "parity/bertopic_gold.py"),
     "Top2Vec": _i("src/top2vec.rs", "src/python/embedding_cluster.rs", "embedding clustering (cluster.rs, reduce.rs)", "embeddings", "parity/top2vec_gold.py, parity/top2vec_compare.py"),
-    "ETM": _i("src/etm.rs", "src/python/mod.rs", "variational EM over word embeddings (ctm.rs)", "", "parity/etm_gold.py"),
+    "ETM": _i("src/etm.rs", "src/python/neural.rs", "variational EM over word embeddings (ctm.rs)", "", "parity/etm_gold.py"),
     "IdealPointTM": _i("src/idealpoint.rs", "src/python/idealpoint.rs", "variational EM + ideal-point head", "", "tests/test_idealpoint.py, tests/test_idealpoint_counts.py"),
     "Wordfish": _i("src/wordfish.rs", "src/python/wordfish.rs", "Poisson-scaling EM", "", "parity/wordfish_r_compare.py, tests/test_wordfish.py"),
     "IdealPointSentenceTM": _i("src/sentence_ideal.rs", "src/python/sentence_ideal.rs", "Gaussian-cluster EM over embeddings", "", "tests/test_sentence_ideal.py"),
@@ -291,9 +291,9 @@ IMPL: dict[str, ImplInfo] = {
     "PartyEmbeddings": _i("src/party_embeddings.rs", "src/python/party_embeddings.rs", "PV-DM paragraph vectors (negative sampling)", "", "parity/party_embeddings_compare.py, tests/test_party_embeddings.py"),
     "FASTopic": _i("src/fastopic.rs", "src/python/mod.rs", "reverse-mode Sinkhorn optimal transport", "", "parity/fastopic_gold.py, parity/fastopic_compare.py"),
     "EmbeddingLDA": _i("python/topica/embedding.py", "", "seeded Gibbs + embedding NN expansion (Python)", "", "parity/embeddinglda_gold.py, tests/test_embedding_lda.py"),
-    "CombinedTM": _i("src/prodlda.rs", "src/python/mod.rs", "contextualized ProdLDA VAE (prodlda.rs)", "", "parity/combinedtm_gold.py, parity/combinedtm_compare.py"),
-    "ZeroShotTM": _i("src/prodlda.rs", "src/python/mod.rs", "contextualized ProdLDA VAE (prodlda.rs)", "", "parity/zeroshot_gold.py, parity/zeroshot_compare.py"),
-    "InfoCTM": _i("src/infoctm.rs", "src/python/mod.rs", "two ProdLDA VAEs + TAMI alignment (prodlda.rs)", "", "parity/infoctm_gold.py, parity/infoctm_compare.py"),
+    "CombinedTM": _i("src/prodlda.rs", "src/python/neural.rs", "contextualized ProdLDA VAE (prodlda.rs)", "", "parity/combinedtm_gold.py, parity/combinedtm_compare.py"),
+    "ZeroShotTM": _i("src/prodlda.rs", "src/python/neural.rs", "contextualized ProdLDA VAE (prodlda.rs)", "", "parity/zeroshot_gold.py, parity/zeroshot_compare.py"),
+    "InfoCTM": _i("src/infoctm.rs", "src/python/neural.rs", "two ProdLDA VAEs + TAMI alignment (prodlda.rs)", "", "parity/infoctm_gold.py, parity/infoctm_compare.py"),
     "TopicGPT": _i("python/topica/llm.py", "", "LLM prompting pipeline (Python)", "", "tests/test_topicgpt.py"),
 }
